Add rate limit for login attempts
Add new database table, php repository and model for a request.
The model takes an ip and sets the time of creation, this model is then inserted into the database.
When another login is attempted the application queries the database for every request with the same ip and within some time frame, if above some threshold then the request is denied and information is given to the user.
Variables for the time frame and maximum number of requests per time frame are stored in the app object.